[#664] Assume node is visible when displayCheck=full if node not in document (#665)
Fixes #664

There are legitimate cases where tabbable's APIs are expected to work on container
nodes that are not actually attached to the document. The result is that the
`getClientRects()` API never returns any rects for visible nodes, resulting in
tabbable thinking ALL nodes in the container are hidden.

__NOTE:__ Assuming the node is visible under the default/legacy "full" mode
when the node is __not__ in the document is actually the legacy behavior.
When a node is detached, `getComputedStyle(node).display` is always
`undefined` and our checks were explicitly verifying if
`getComputedStyle(node).display === 'none'` to conclude a node was
hidden, which means this check would fail and we could conclude the
node was __visible__ even if it was not.