Migrate to Cypress 10
Ran `cypress open` and walked through migration guide. It generated cypress.config.ts
which I renamed to cypress.config.js and reverted to `require()` instead of `import`
because of stupid TypeScript typings warnings related to the plugins, which seem
perfectly fine.

e2e tests all run and pass as before.

Unfortunately, coverage is broken because we also need
`@cypress/code-coverage@3.10.0` which supports Cypress 10, but
it has an issue with this error:
https://github.com/cypress-io/code-coverage/issues/573