[832] Fix bug with tabbable scoped table headers
Fixes #832

For shadow DOM support, we create internal objects with a `scope` property
when looking for tabbable/focusable candidates. But when we sort the
tabbable candidates, we look for an existing `scope` property assuming
that HTMLElements don't have one, but it turns out `<th>` elements do
have a `scope` attribute (and so `scope` property) and we mistake it
for being a shadow DOM scope parent element. Crashing ensues.

So this just renames `CandidateScope.scope` to `CandidateScope.scopeParent`
and adds a test to make sure things work as expected.