Remove sorting of `containers`
It's not necessary as discussed here https://github.com/focus-trap/focus-trap/issues/375#issuecomment-1591807907:

If tabbable starts to resort containers, it will probably break expectations/behavior in focus-trap. There's a now long-standing implication that when multiple containers are given, they're given in the order in which focus should move between them, not a random order for focus-trap (and now tabbable) to sort out make "proper".