Fix tiny bug in tab index parsing
Even though in practice `Element` have a `tabIndex` property, according to the specs it doesn't.

So the following check is needed in `getTabIndex`:

```
if (!('tabIndex' in node && typeof node.tabIndex === 'number')) {
  return something
}
```

But what should be the returned `something`?
Adding that check inside `getTabIndex` would inevitably add overall complexity in the library.

Since `getTabIndex` is not pubblic API, I believe it would be reasonable to just document that it expects an `Element` with a numeric `tabIndex` property.
So that the check can be introduced elsewhere.

`isNodeMatchingSelectorTabbable` is the ideal place for the check, because it's where the tab index analysis of every element starts and it's also already  a filter for (among other things) elements with a certain tab index.