feat: add 'full-native' `displayCheck` option, utilising Element#checkVisibility (#1818)
This PR closes #1800, by introducing a 'full-native' displayCheck option. The option is done as a progressive enhancement on top of 'full', meaning that if Element#checkVisibility is supported, then it is used and its value returned verbatim. Otherwise, we follow the existing codepaths with manual checks, that resemble the browser behavior.

I parameterised the tests tests to include this option, and they pass the same as before, which gives me high confidence (very extensive test suite btw 😌).

* Fix setupTestWindow helper

* Add checkVisibility branch to 'full' method

* Fast-path checkVisibility, over getComputedStyle and details/summary

* Add 'full-native' checkVisibility option

* Add types

* Add changeset

* Align option name in README

* Remove done TODO

* Address fallthrough case in manual checks

* Fix firefox inertness exclusion for CI

* Add content-visibility: hidden test cases

* Clarify native behaviour of opacity: 0

* Reorder full-native option in README, remove line about defaults

* Reorder full-native option in types

* Unit test full-native fallback to full

* Account for Firefox < 125 checkVisiblity with content-visibility: hidden parent in tests

* Document choice to exclude opacity checks

* Fix test condition to filter either on !== 'full-native' or isFirefoxLowerThan125()

* Fix firefox detection in tests (browser.name is lowercase)